This commit focuses on section 2.0 §2.0¶1 The opening of §2.0¶1 is one of the key sections to which readers new to HQ/QW will be drawn when first skimming the SRD. As such, I feel it should be stronger, and introduce the key concepts.
§2.0¶2
While conflict vs task based resolution is addressed in some detail
further down, this top section is again where  readers will get their
first impression of the rules. Such a key point should be introduced
here, even if in brief.

Also, I replace "success"/"failure" here with "victory"/"defeat."
The distinction between "Success vs Failure" and "Victory vs Defeat"
is one of the subtleties that trip people up.

As such, S/F should only and always be used to describe results of
ability or resistance rolls, while conversely V/D should conversely be
only/always used when referring to the results of a conflict.

The above guidance may even be worth explicitly including in the SRD,
as it's a distinction that even veteran designers can easily mix up in
passing (case in point :) ), It would be good to point it out explicitly
to both players/GMs and implementers alike.

§2.0¶3
Eliminated dupe word.
Changed "changed "frame the conflict" to "frame the contest" to
match the way the term is presented elsewhere.

§2.0¶5
Changed plural "dice" to singular "die" to agree with rest of sentance.
Reworded to highlight the Success/Failure-Victory/Defeat distinction.

§2.0¶6
Changed "success or failure" to victory or defeat.