This commit addresses §2.3 - §2.3.2.2 Obstacles. I dig quite a bit here into wording surrounding terminology and its application to examples. In particular, the increased  focus on the "Obstacle" terminology relative to previous, while good on the whole has in this section obscured some other concepts and terms.
§2.3
Changed name to "Contest Proceedure"
"Contest" should be preferred for the nuts and bolts of die-rolling, as
distinct from references to conflicts with obstacles.
Grammar fix: "describes" to "describe"

§2.3.1.1
Changed name to Automatic Victory, as it's actually talking about the
conflict results rather than the contest results.
Reworded for consistant terminology.
Reworded to clarify reasons for application.
Removed part of the find another way bit, it's already addressed under
abilities above, ans this sort of conflicted with that.
Added reference to related optional rules sections.

§2.3.1.2
Added elaboration to introduce the significance of the simple contest
in the context of the game, and provide a bit more of an overview.
Smoothed out the PvP wording.

§2.3.2.1
Added context and bullet point for goal.

§2.3.2.2
Changed name to "Conflict: Goals vs Obstacles" - continuing to drive
home the terms and their relationships at every opportunity.
Changed 'succeeded' to 'performed.'
Many changes to the examples, restoring goal verbiage alongside
obstacle verbiage, and to avoud goal/means ambiguity.
Switched rescuing sister to father, because stereotypes.